refactor: move WP_Google_Places_Reviews_Free into separate file
